Separates elements of the dashboard in widgets
