+++ lib/pages/dashboard/widgets/dashboard_club_teams.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -10,11 +8,9 @@ import 'package:v34/models/club.dart';
 import 'package:v34/models/team.dart';
 import 'package:v34/pages/club-details/blocs/club_teams.bloc.dart';
 import 'package:v34/pages/club-details/club_detail_page.dart';
-import 'package:v34/pages/dashboard/team_card.dart';
+import 'package:v34/pages/dashboard/widgets/team_card.dart';
 import 'package:v34/pages/team-details/team_detail_page.dart';
 import 'package:v34/repositories/repository.dart';
-
-final Random random = Random.secure();
 
 typedef TeamFavoriteChangeCallback = void Function(Team team);
 
@@ -137,7 +133,4 @@ class _DashboardClubTeamsState extends State<DashboardClubTeams> with SingleTick
       },
     );
   }
-
-  @override
-  bool get wantKeepAlive => true;
 }
